(Refactor) Extract ActionLine tests to their own file
