+++ Assets/Tests/EditModeTests/ActionLineTests.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: dce21ac5f60c4fc4db97e3295d1a789c
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
